Implement `cancel_p2p_swap`

+++ backend/integration_tests/src/p2p_swap_tests.rs
@@ -6,6 +6,7 @@ use candid::Principal;
 use pocket_ic::PocketIc;
 use std::ops::Deref;
 use std::time::Duration;
+use test_case::test_case;
 use types::{ChatEvent, ChatId, Cryptocurrency, MessageContent, MessageContentInitial, P2PSwapContentInitial, P2PSwapStatus};
 use utils::time::{DAY_IN_MS, MINUTE_IN_MS};
 
@@ -224,8 +225,9 @@ fn p2p_swap_in_group_succeeds() {
     }
 }
 
-#[test]
-fn cancel_p2p_swap_in_direct_chat_succeeds() {
+#[test_case(true)]
+#[test_case(false)]
+fn cancel_p2p_swap_in_direct_chat_succeeds(delete_message: bool) {
     let mut wrapper = ENV.deref().get();
     let TestEnv {
         env,
@@ -276,24 +278,42 @@ fn cancel_p2p_swap_in_direct_chat_succeeds() {
         user_canister::send_message_v2::Response::TransferSuccessV2(_)
     ));
 
-    let delete_message_response = client::user::delete_messages(
-        env,
-        user1.principal,
-        user1.canister(),
-        &user_canister::delete_messages::Args {
-            user_id: user2.user_id,
-            thread_root_message_index: None,
-            message_ids: vec![message_id],
-            correlation_id: 0,
-        },
-    );
+    if delete_message {
+        let delete_message_response = client::user::delete_messages(
+            env,
+            user1.principal,
+            user1.canister(),
+            &user_canister::delete_messages::Args {
+                user_id: user2.user_id,
+                thread_root_message_index: None,
+                message_ids: vec![message_id],
+                correlation_id: 0,
+            },
+        );
 
-    assert!(matches!(
-        delete_message_response,
-        user_canister::delete_messages::Response::Success
-    ));
+        assert!(matches!(
+            delete_message_response,
+            user_canister::delete_messages::Response::Success
+        ));
 
-    env.advance_time(Duration::from_millis(10 * MINUTE_IN_MS));
+        env.advance_time(Duration::from_millis(5 * MINUTE_IN_MS));
+    } else {
+        let cancel_swap_response = client::user::cancel_p2p_swap(
+            env,
+            user1.principal,
+            user1.canister(),
+            &user_canister::cancel_p2p_swap::Args {
+                user_id: user2.user_id,
+                message_id,
+            },
+        );
+
+        assert!(matches!(
+            cancel_swap_response,
+            user_canister::cancel_p2p_swap::Response::Success
+        ));
+    }
+
     tick_many(env, 5);
 
     assert_eq!(
